+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 - fixed bug in `isIPV4` - no longer allows leading zeros
 - improved implementation of `isJSONPointer` and `isRelativeJSONPointer`
+- added `isRegex`
 
 ## [1.3] - 2021-02-25
 ### Changed